// src/enums/keyword.ts
// cSpell:disable

enum Keyword {
  BaseDependent = "base",
  Brief = "bref",
  Changed = "changed",
  ContinuedFraction = "cofr",
  DecimalExpansion = "cons",
  Core = "core",
  Dead = "dead",
  Unimportant = "dumb",
  Duplicate = "dupe",
  Easy = "easy",
  EigenSequence = "eigen",
  Finite = "fini",
  Fraction = "frac",
  Full = "full",
  Hard = "hard",
  Audible = "hear",
  LessInteresting = "less",
  VisualInterest = "look",
  MoreTermsNeeded = "more",
  Multiplicative = "mult",
  New = "new",
  Nice = "nice",
  Nonnegative = "nonn",
  Obscure = "obsc",
  Probationary = "probation",
  Signed = "sign",
  IrregularTriangle = "tabf",
  RegularTriangle = "tabl",
  Unedited = "uned",
  Unknown = "unkn",
  WalkCounting = "walk",
  WordBased = "word",
}

export default Keyword;